Eliminate integer static cast

+++ Common/DataModel/vtkQuadratureSchemeDefinition.cxx
@@ -354,7 +354,7 @@ int vtkQuadratureSchemeDefinition::SaveState(vtkXMLDataElement* root)
       ssShapeWts << " " << this->ShapeFunctionWeights[id];
     }
     string sShapeWts = ssShapeWts.str();
-    eShapeWts->SetCharacterData(sShapeWts.c_str(), static_cast<int>(sShapeWts.size()));
+    eShapeWts->SetCharacterData(sShapeWts.c_str(), sShapeWts.size());
 
     // Write quadrature weights
     ostringstream ssQuadWts;
@@ -366,7 +366,7 @@ int vtkQuadratureSchemeDefinition::SaveState(vtkXMLDataElement* root)
       ssQuadWts << " " << this->QuadratureWeights[id];
     }
     string sQuadWts = ssQuadWts.str();
-    eQuadWts->SetCharacterData(sQuadWts.c_str(), static_cast<int>(sQuadWts.size()));
+    eQuadWts->SetCharacterData(sQuadWts.c_str(), sQuadWts.size());
   }
   else
   {
